requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-JWT-Extended
marshmallow
requests
python-dotenv
fpdf
psycopg2-binary
langdetect
matplotlib
plotly
numpy
vaderSentiment
Werkzeug
python-dateutil
pytz
gunicorn
python-docx 
flask-migrate